Add Source type
