+++ denops/skkeleton/dictionary.ts
@@ -3,11 +3,11 @@ import { JpNum } from "./deps/japanese_numeral.ts";
 import { RomanNum } from "./deps/roman.ts";
 import { zip } from "./deps/std/collections.ts";
 import type { CompletionData, RankData } from "./types.ts";
-import { SkkDictionary } from "./sources/skk_dictionary.ts";
-import { DenoKvDictionary } from "./sources/deno_kv.ts";
 import { UserDictionary } from "./sources/user_dictionary.ts";
-import { SkkServer } from "./sources/skk_server.ts";
-import { GoogleJapaneseInput } from "./sources/google_japanese_input.ts";
+import { SkkDictionarySource } from "./sources/skk_dictionary.ts";
+import { DenoKvSource } from "./sources/deno_kv.ts";
+import { SkkServerSource } from "./sources/skk_server.ts";
+import { GoogleJapaneseInputSource } from "./sources/google_japanese_input.ts";
 
 export const okuriAriMarker = ";; okuri-ari entries.";
 export const okuriNasiMarker = ";; okuri-nasi entries.";
@@ -111,6 +111,10 @@ function convertNumber(pattern: string, entry: string): string {
 export interface Dictionary {
   getHenkanResult(type: HenkanType, word: string): Promise<string[]>;
   getCompletionResult(prefix: string, feed: string): Promise<CompletionData>;
+}
+
+export interface Source {
+  getDictionaries(): Promise<Dictionary[]>;
 }
 
 export class NumberConvertWrapper implements Dictionary {
@@ -271,74 +275,24 @@ export async function load(): Promise<Library> {
     // do nothing
   }
 
-  const dictionaries: Dictionary[] = [];
+  let dictionaries: Dictionary[] = [];
   for (const source of config.sources) {
     if (source === "skk_dictionary") {
-      const globalDictionaries = await Promise.all(
-        config.globalDictionaries.map(async ([path, encodingName]) => {
-          try {
-            const dict = new SkkDictionary();
-            await dict.load(path, encodingName);
-            return dict;
-          } catch (e) {
-            console.error("globalDictionary loading failed");
-            console.error(`at ${path}`);
-            if (config.debug) {
-              console.error(e);
-            }
-            return undefined;
-          }
-        }),
+      dictionaries = dictionaries.concat(
+        await (new SkkDictionarySource().getDictionaries()),
       );
-
-      for (const d of globalDictionaries) {
-        if (d) {
-          dictionaries.push(wrapDictionary(d));
-        }
-      }
     } else if (source === "deno_kv") {
-      const globalDictionaries = await Promise.all(
-        config.globalDictionaries.map(async ([path, encodingName]) => {
-          try {
-            const dict = await DenoKvDictionary.create(path, encodingName);
-            await dict.load();
-            return dict;
-          } catch (e) {
-            console.error("globalDictionary loading failed");
-            console.error(`at ${path}`);
-            if (config.debug) {
-              console.error(e);
-            }
-            return undefined;
-          }
-        }),
+      dictionaries = dictionaries.concat(
+        await (new DenoKvSource().getDictionaries()),
       );
-
-      for (const d of globalDictionaries) {
-        if (d) {
-          dictionaries.push(wrapDictionary(d));
-        }
-      }
     } else if (source === "skk_server") {
-      const skkServer = new SkkServer({
-        hostname: config.skkServerHost,
-        port: config.skkServerPort,
-        requestEnc: config.skkServerReqEnc,
-        responseEnc: config.skkServerResEnc,
-      });
-
-      try {
-        skkServer.connect();
-      } catch (e) {
-        if (config.debug) {
-          console.log("connecting to skk server is failed");
-          console.log(e);
-        }
-      }
-
-      dictionaries.push(skkServer);
+      dictionaries = dictionaries.concat(
+        await (new SkkServerSource().getDictionaries()),
+      );
     } else if (source === "google_japanese_input") {
-      dictionaries.push(new GoogleJapaneseInput());
+      dictionaries = dictionaries.concat(
+        await (new GoogleJapaneseInputSource().getDictionaries()),
+      );
     } else {
       console.error(`Invalid source name: ${source}`);
     }
